fix: update internal imports in moved files to use new paths
