+++ shargain/offers/field_extraction/plugins/core_fields.py
@@ -1,0 +1,29 @@
+"""Core Fields Plugin -- applies to all URLs, extracts basic offer fields."""
+
+from shargain.offers.models import Offer
+from shargain.offers.field_extraction.plugin import (
+    BaseFieldPlugin,
+    FieldDefinition,
+    FieldType,
+    ListUrl,
+)
+
+
+class CoreFieldsPlugin(BaseFieldPlugin):
+    """Provides 'title' and 'price' for all offers. Registered first for backward compat."""
+
+    def matches(self, url: ListUrl) -> bool:
+        return True
+
+    @property
+    def fields(self) -> list[FieldDefinition]:
+        return [
+            FieldDefinition("title", "Title", FieldType.STRING),
+            FieldDefinition("price", "Price", FieldType.NUMBER, unit="z\u0142"),
+        ]
+
+    def extract(self, offer: Offer, url: ListUrl) -> dict:
+        return {"title": offer.title, "price": offer.price}
+
+
+core_fields = CoreFieldsPlugin()
